--- js/data.js
@@ -1,173 +1,0 @@
-import {getRandomIntInclusive} from './util.js';
-
-/*на основе написанных по заданию ранее вспомогательных функций напишите необходимые функции для создания массива из 25 сгенерированных объектов. Каждый объект массива — описание фотографии, опубликованной пользователем.
-
-Структура каждого объекта должна быть следующей:
-
-id, число — идентификатор описания. Это число от 1 до 25. Идентификаторы не должны повторяться.
-
-url, строка — адрес картинки вида photos/{{i}}.jpg, где {{i}} — это число от 1 до 25. Адреса картинок не должны повторяться.
-
-description, строка — описание фотографии. Описание придумайте самостоятельно.
-
-likes, число — количество лайков, поставленных фотографии. Случайное число от 15 до 200.
-
-comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии. Количество комментариев к каждой фотографии вы определяете на своё усмотрение. Все комментарии генерируются случайным образом. Пример описания объекта с комментарием:
-
-{
-  id: 135,
-  avatar: 'img/avatar-6.svg',
-  message: 'В целом всё неплохо. Но не всё.',
-  name: 'Артём',
-}
-У каждого комментария есть идентификатор — id — случайное число. Идентификаторы не должны повторяться.
-
-Поле avatar — это строка, значение которой формируется по правилу img/avatar-{{случайное число от 1 до 6}}.svg. Аватарки подготовлены в директории img.
-
-Для формирования текста комментария — message — вам необходимо взять одно или два случайных предложения из представленных ниже:
-
-Всё отлично!
-В целом всё неплохо. Но не всё.
-Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
-Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
-Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
-Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!
-Имена авторов также должны быть случайными. Набор имён для комментаторов составьте сами. Подставляйте случайное имя в поле name.*/
-
-//выносим кол-во фото в переменную, чтобы не плодить магические числа
-const PHOTO_COUNT = 25;
-//создаем массив с комментами
-const COMMENTS = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-//создаем массив с именами комментаторов
-const NAMES = [
-  'Артем',
-  'Иннокентий Викторович',
-  'Василий',
-  'Иннусик',
-  'Энни Лейбовиц',
-  'Руслан В.',
-  'Клара Георгиевна',
-  'Петро',
-  'Арам',
-  'Хельмут Ньютон',
-  'Остап',
-  'Майя',
-  'Алина Абрамович',
-  'Митрофан',
-  'Ангелина',
-  'Виталий',
-];
-//создаем массив с описаниями фото
-const DESCRIPTIONS = [
-  'Восемь котов',
-  'Фетуччини с лососем',
-  'Я на курорте',
-  'Закат на районе',
-  'Думаю о вечном',
-  'Любимый сфоткал, пока я спала',
-  'Вкалываю на тренировке',
-  'Рука об руку в будущее',
-  'Я читаю Коэльо',
-  'Побывал на выставке',
-  'Стою возле крутой тачки',
-  'Поймал леща',
-  'Девичник в белом кролике',
-  'Розовый младенец',
-  'Задумчиво смотрю вдаль',
-  'Стою рядом с девушками',
-  'Любимый подарил',
-  'Шашлыки в Бутово',
-  'Селфи в лифте в деловом костюме',
-  'Панорама Химкинского водохранилища',
-  'Сумка, мобильник, духи, брелок от авто',
-  'Умею кататься на вейке',
-  'Составлю натальную карту',
-  'Записывайтесь ко мне на брови',
-  'Соберу портфель инвестиций',
-];
-
-//создаем пустой массив с id фоток, добавляем туда 25 элементов с числами от 1 до 25
-const ids = [];
-for (let i = 0; i < 25; i++){
-  ids.push(i + 1);
-}
-
-//создаем пустой массив с id комментариев, добавляем туда 200 элементов с числами от 0 до 199
-const commentIds = [];
-for (let i = 0; i < 200; i++){
-  commentIds.push(i);
-}
-
-/**
- * Берем из массива случайный элемент и тут же удаляем его из массива
- *
- * @param {object} array - массив
- */
-const getRandomElementFromArray = (array) => array.splice(getRandomIntInclusive(0, array.length-1), 1);
-
-/**
- * Создаем объект с комментарием, его id, аватаркой комментатора, текстом коммента и именем комментатора
- *
- * @returns {object}
- */
-const createComment = () => {
-  let commentId = getRandomElementFromArray(commentIds);
-  commentId = commentId[0];
-  const randomNameIndex = getRandomIntInclusive(0, NAMES.length - 1);
-  const randomCommentIndex = getRandomIntInclusive(0, COMMENTS.length - 1);
-  return {
-    id: commentId,
-    avatar: `img/avatar-${getRandomIntInclusive(0, 6)}.svg`,
-    message: COMMENTS[randomCommentIndex],
-    commenterName: NAMES[randomNameIndex],
-  };
-};
-
-/**
- * Создаем массив из объектов с комментариями
- *
- * @returns {object}
- */
-const createComments = () => {
-  const commentsNum = getRandomIntInclusive(0, 5);
-  const comments = [];
-  for (let i = 0; i < commentsNum; i++) {
-    comments.push(createComment());
-  }
-  return comments;
-};
-
-/**
- * Создаем объект с инфой о фото, включающий id фото, адрес, описание, кол-во лайков и комментарии
- *
- * @returns {object}
- */
-const createPhotoDescription = () => {
-  let  id = getRandomElementFromArray(ids);
-  id = id[0];
-  const randomDescriptionIndex = getRandomIntInclusive(0, DESCRIPTIONS.length - 1);
-  return {
-    id: id,
-    url: `photos/${id}.jpg`,
-    description: DESCRIPTIONS[randomDescriptionIndex],
-    likes: getRandomIntInclusive(15, 200),
-    comments: createComments(),
-  };
-};
-
-/**
- * Создаем массив из объектов с инфой о фото
- *
- * @returns {object}
- */
-const photos = new Array(PHOTO_COUNT).fill(null).map(() => createPhotoDescription());
-
-// eslint-disable-next-line no-console
-console.log(photos);
