+++ js/mock/data.js
@@ -1,0 +1,61 @@
+//создаем массив с комментами
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+//создаем массив с именами комментаторов
+const NAMES = [
+  'Артем',
+  'Иннокентий Викторович',
+  'Василий',
+  'Иннусик',
+  'Энни Лейбовиц',
+  'Руслан В.',
+  'Клара Георгиевна',
+  'Петро',
+  'Арам',
+  'Хельмут Ньютон',
+  'Остап',
+  'Майя',
+  'Алина Абрамович',
+  'Митрофан',
+  'Ангелина',
+  'Виталий',
+];
+//создаем массив с описаниями фото
+const DESCRIPTIONS = [
+  'Восемь котов',
+  'Фетуччини с лососем',
+  'Я на курорте',
+  'Закат на районе',
+  'Думаю о вечном',
+  'Любимый сфоткал, пока я спала',
+  'Вкалываю на тренировке',
+  'Рука об руку в будущее',
+  'Я читаю Коэльо',
+  'Побывал на выставке',
+  'Стою возле крутой тачки',
+  'Поймал леща',
+  'Девичник в белом кролике',
+  'Розовый младенец',
+  'Задумчиво смотрю вдаль',
+  'Стою рядом с девушками',
+  'Любимый подарил',
+  'Шашлыки в Бутово',
+  'Селфи в лифте в деловом костюме',
+  'Панорама Химкинского водохранилища',
+  'Сумка, мобильник, духи, брелок от авто',
+  'Умею кататься на вейке',
+  'Составлю натальную карту',
+  'Записывайтесь ко мне на брови',
+  'Соберу портфель инвестиций',
+];
+
+export { COMMENTS, NAMES, DESCRIPTIONS };
+
+
