+++ js/utils/check-string-length.js
@@ -1,3 +1,6 @@
 function checkStringLength (string, length) {
   return string.length <= length;
 }
+
+// eslint-disable-next-line no-console
+console.log(checkStringLength(10, 25));
